requirements.txt
torch~=1.13.0+cu116
torchvision~=0.14.0+cu116
tqdm~=4.64.1
scikit-image~=0.19.3
pillow~=9.3.0